% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_elo_predictions.R
\name{adjust_for_phase}
\alias{adjust_for_phase}
\title{Adjust for Match Phase}
\usage{
adjust_for_phase(expected_runs, phase, match_type)
}
\description{
Adjust for Match Phase
}
\keyword{internal}
