% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_elo_predictions.R
\name{adjust_wicket_for_over}
\alias{adjust_wicket_for_over}
\title{Adjust Wicket Probability for Over}
\usage{
adjust_wicket_for_over(wicket_prob, over_num, match_type)
}
\description{
Adjust Wicket Probability for Over
}
\keyword{internal}
