% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_elo_predictions.R
\name{adjust_wicket_for_phase}
\alias{adjust_wicket_for_phase}
\title{Adjust Wicket Probability for Phase}
\usage{
adjust_wicket_for_phase(wicket_prob, phase, match_type)
}
\description{
Adjust Wicket Probability for Phase
}
\keyword{internal}
