% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{aggregate_match_results}
\alias{aggregate_match_results}
\title{Aggregate Match Simulation Results}
\usage{
aggregate_match_results(results, team1, team2)
}
\arguments{
\item{results}{List. List of simulation results from simulate_match_outcome}

\item{team1}{Character. Team 1 name}

\item{team2}{Character. Team 2 name}
}
\value{
List with aggregated statistics
}
\description{
Aggregates results from multiple match simulations.
}
