% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\docType{data}
\name{ALL_MATCH_TYPES}
\alias{ALL_MATCH_TYPES}
\title{All Valid Match Types}
\format{
An object of class \code{character} of length 6.
}
\usage{
ALL_MATCH_TYPES
}
\description{
All Valid Match Types
}
\keyword{internal}
