% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality_storage.R
\name{batch_get_metric_for_match}
\alias{batch_get_metric_for_match}
\title{Batch get metric values for a match}
\usage{
batch_get_metric_for_match(
  player_ids,
  match_date,
  metric,
  format,
  conn,
  gender = NULL
)
}
\description{
Batch get metric values for a match
}
\keyword{internal}
