% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostingControl.R
\name{boostingControl}
\alias{boostingControl}
\title{Control Function for the Boosting Setup}
\usage{
boostingControl(mstop = 100, nu = 0.1, base = "lm")
}
\arguments{
\item{mstop}{an integer specifying the number of boosting iterations for each weak learner}

\item{nu}{a numeric between 0 and 1 specifying the learning rate for each weak learner}

\item{base}{a character specifying the type of baslearner to use. One of "lm" or "tree", either
uses linear baselearners or tree baselearner.}
}
\value{
a list of the specified input
}
\description{
Control Function for the Boosting Setup
}
\examples{
boosting <- boostingControl(mstop = 1000, nu = 0.05)

}
