% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_processing.R
\name{build_delivery_query}
\alias{build_delivery_query}
\title{Build Delivery Query}
\usage{
build_delivery_query(conn, matches_with_elo, match_limit)
}
\description{
Build Delivery Query
}
\keyword{internal}
