% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_elo_format_tables.R
\name{build_elo_params}
\alias{build_elo_params}
\title{Build Current ELO Parameters}
\usage{
build_elo_params(format = "t20")
}
\arguments{
\item{format}{Character. Format: "t20", "odi", or "test"}
}
\value{
Named list of parameters
}
\description{
Creates a list of current ELO parameters from constants.
}
\keyword{internal}
