% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_utils.R
\name{build_match_type_sql}
\alias{build_match_type_sql}
\title{Build SQL Match Type Filter}
\usage{
build_match_type_sql(format, column = "match_type")
}
\arguments{
\item{format}{Character. Canonical format or match type.}

\item{column}{Character. Column name to filter on. Default "match_type".}
}
\value{
Character. SQL fragment like "match_type IN ('T20', 'IT20')"
}
\description{
Generates SQL WHERE clause fragment for filtering by format.
}
\examples{
build_match_type_sql("t20")
# Returns: "match_type IN ('T20', 'IT20')"

build_match_type_sql("test", "m.match_type")
# Returns: "m.match_type IN ('Test', 'MDM')"

}
