% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{builder}
\alias{builder}
\title{Build formulas out of an output returned by \code{\link{featureSelection}}
This function allows you to build custom estimation functions from a \linkS4class{sel_obj}.}
\usage{
builder(object, n_features = 5)
}
\arguments{
\item{object}{must be an object of class \linkS4class{sel_obj}}

\item{n_features}{a \code{\link{numeric}} value indicating the number of features to add into the equation}
}
\value{
A model formula suggested by the feature selection algorithm.
}
\description{
Build formulas out of an output returned by \code{\link{featureSelection}}
This function allows you to build custom estimation functions from a \linkS4class{sel_obj}.
}
\examples{
\dontrun{
# simulate data
test_df <- sim_data()

# feature selection
sel <- featureSelection(data = test_df,
                        target = "y",
                        index = NULL,
                        n_cores = 1)
# extract one feature
form <- builder(object = sel, n_features = 1)
}
}
