% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_match_prediction.R
\name{calculate_innings1_win_prob}
\alias{calculate_innings1_win_prob}
\title{Calculate Chase Win Probability from Model}
\usage{
calculate_innings1_win_prob(data, model, feature_cols)
}
\arguments{
\item{data}{List with feature data (projected_final_score, etc.)}

\item{model}{XGBoost model for innings 1 win probability}

\item{feature_cols}{Character vector of feature column names}
}
\value{
Numeric win probability (0-1) for chasing team

Numeric win probability (0-1)
}
\description{
Internal function to get chase win probability from Stage 2 model.
}
\keyword{1}
\keyword{Calculate}
\keyword{Innings}
\keyword{Model}
\keyword{Probability}
\keyword{Trained}
\keyword{Using}
\keyword{Win}
\keyword{internal}
