% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_core.R
\name{calculate_k_factor}
\alias{calculate_k_factor}
\title{Calculate K-Factor}
\usage{
calculate_k_factor(match_type, player_matches = 0)
}
\arguments{
\item{match_type}{Character. Type of match ("test", "odi", "t20", etc.)}

\item{player_matches}{Numeric. Number of matches player has played}
}
\value{
Numeric K-factor value
}
\description{
Calculates the K-factor (learning rate) for ELO updates based on match type
and player experience.
}
\examples{
# New player in Test match
calculate_k_factor("test", player_matches = 5)

# Experienced player in T20
calculate_k_factor("t20", player_matches = 100)
}
