% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_way_glicko.R
\name{calculate_mae}
\alias{calculate_mae}
\title{Calculate MAE}
\usage{
calculate_mae(actual, expected)
}
\arguments{
\item{actual}{Numeric vector. Actual outcomes.}

\item{expected}{Numeric vector. Expected values.}
}
\value{
Numeric. MAE.
}
\description{
Mean Absolute Error for comparison with other metrics.
}
\keyword{internal}
