% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_match_prediction.R
\name{calculate_projected_score_from_model}
\alias{calculate_projected_score_from_model}
\title{Calculate Projected Score from Model}
\usage{
calculate_projected_score_from_model(data, model, feature_cols, format)
}
\arguments{
\item{data}{data.frame with feature data}

\item{model}{XGBoost model}

\item{feature_cols}{Feature column names}

\item{format}{Match format}
}
\value{
Numeric projected score
}
\description{
Internal function to get projected score from Stage 1 model.
}
\keyword{internal}
