% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_way_glicko.R
\name{calculate_rmse}
\alias{calculate_rmse}
\title{Calculate RMSE}
\usage{
calculate_rmse(actual, expected)
}
\arguments{
\item{actual}{Numeric vector. Actual outcomes.}

\item{expected}{Numeric vector. Expected values.}
}
\value{
Numeric. RMSE.
}
\description{
Root Mean Squared Error for comparison with other metrics.
}
\keyword{internal}
