% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cricsheet_parser.R
\name{classify_match}
\alias{classify_match}
\title{Classify Match for Data Folder Organization}
\usage{
classify_match(match_info)
}
\arguments{
\item{match_info}{Data frame or list with match information.
Must contain: match_type, gender}
}
\value{
Character: folder name like "short_form_male_club"
}
\description{
Generates a folder name for organizing match JSON files.
Combines format (long_form/short_form), gender, and type (international/club).
}
\details{
The 8 possible folder names are:
\itemize{
\item long_form_male_international
\item long_form_male_club
\item long_form_female_international
\item long_form_female_club
\item short_form_male_international
\item short_form_male_club
\item short_form_female_international
\item short_form_female_club
}
}
\examples{
\dontrun{
match_data <- parse_cricsheet_json("12345.json")
folder <- classify_match(match_data$match_info)
# e.g., "short_form_male_club" for an IPL match
}
}
\keyword{internal}
