% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_units.R
\docType{data}
\name{COMPETITION_REDUCED_SEASONS}
\alias{COMPETITION_REDUCED_SEASONS}
\title{Seasons Where a Competition Ran in a Reduced Format}
\format{
An object of class \code{list} of length 1.
}
\usage{
COMPETITION_REDUCED_SEASONS
}
\description{
Strength estimates for these are less comparable to the competition's other seasons.
Report them, do not silently drop them.
}
\keyword{datasets}
