% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_rating_v2.R
\docType{data}
\name{COMPETITION_REFERENCE_T20}
\alias{COMPETITION_REFERENCE_T20}
\title{Reference Competitions Defining the 1.0 Difficulty Scale}
\format{
An object of class \code{character} of length 9.
}
\usage{
COMPETITION_REFERENCE_T20
}
\description{
Reference Competitions Defining the 1.0 Difficulty Scale
}
\keyword{datasets}
