% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delivery_simulation.R
\name{create_default_batters}
\alias{create_default_batters}
\title{Create Default Batter Skills}
\usage{
create_default_batters(n = 11, format = "t20")
}
\arguments{
\item{n}{Integer. Number of batters to create (default 11)}

\item{format}{Character. Format for default values}
}
\value{
List of batter skill objects
}
\description{
Creates a list of default batter skill objects for simulation.
}
\keyword{internal}
