% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delivery_simulation.R
\name{create_default_bowlers}
\alias{create_default_bowlers}
\title{Create Default Bowler Skills}
\usage{
create_default_bowlers(n = 6, format = "t20")
}
\arguments{
\item{n}{Integer. Number of bowlers to create (default 6)}

\item{format}{Character. Format for default values}
}
\value{
List of bowler skill objects
}
\description{
Creates a list of default bowler skill objects for simulation.
}
