% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cricinfo_scraper.R
\name{cricinfo_discover_matches}
\alias{cricinfo_discover_matches}
\title{Discover Matches in a Cricinfo Series}
\usage{
cricinfo_discover_matches(browser, series_id)
}
\arguments{
\item{browser}{A ChromoteSession object}

\item{series_id}{Integer series objectId}
}
\value{
data.frame of matches, or NULL
}
\description{
Navigates to the series schedule page and extracts match listing
from the embedded Next.js data.
}
\examples{
\dontrun{
browser <- chromote::ChromoteSession$new()
matches <- cricinfo_discover_matches(browser, series_id = 1455609)
}

}
