% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cricinfo_scraper.R
\name{cricinfo_series_url}
\alias{cricinfo_series_url}
\title{Build ESPN Cricinfo Series Page URL}
\usage{
cricinfo_series_url(series_id, page = "match-schedule-fixtures-and-results")
}
\arguments{
\item{series_id}{Integer series objectId}

\item{page}{Page type: "match-schedule-fixtures-and-results"}
}
\value{
Character URL
}
\description{
Build ESPN Cricinfo Series Page URL
}
\keyword{internal}
