% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_rating_v2.R
\name{default_competition_reference}
\alias{default_competition_reference}
\title{Default Reference Set for a Bucket}
\usage{
default_competition_reference(format = "t20", gender = "male")
}
\arguments{
\item{format, gender}{Bucket.}
}
\value{
Character vector of competition names.
}
\description{
Default Reference Set for a Bucket
}
