% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_rating_v2.R
\name{default_exposure_floor}
\alias{default_exposure_floor}
\title{The exposure floor for a format and role}
\usage{
default_exposure_floor(format, role)
}
\description{
The exposure floor for a format and role
}
\keyword{internal}
