% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality_storage.R
\name{delete_old_metric_snapshots}
\alias{delete_old_metric_snapshots}
\title{Delete old metric snapshots}
\usage{
delete_old_metric_snapshots(
  metric,
  format,
  keep_months = CENTRALITY_SNAPSHOT_KEEP_MONTHS,
  conn,
  gender = NULL
)
}
\description{
Delete old metric snapshots
}
\keyword{internal}
