% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_match_features.R
\name{detect_knockout_match}
\alias{detect_knockout_match}
\title{Detect Knockout Match}
\usage{
detect_knockout_match(event_match_number, event_group)
}
\arguments{
\item{event_match_number}{Character or Integer. Match number in event}

\item{event_group}{Character. Group/stage information}
}
\value{
Logical. TRUE if knockout match
}
\description{
Determines if a match is a knockout/playoff match based on event info.
}
