% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_game_data.R
\name{.aggregate_batting_game_data}
\alias{.aggregate_batting_game_data}
\title{Aggregate Batting Stats Per Player Per Match}
\usage{
.aggregate_batting_game_data(conn, format, match_ids = NULL, gender = NULL)
}
\arguments{
\item{conn}{DBI connection}

\item{format}{Character. Match format.}

\item{match_ids}{Character vector or NULL.}

\item{gender}{Character or NULL.}
}
\value{
data.table with batting stats per batter per match.
}
\description{
Aggregate Batting Stats Per Player Per Match
}
\keyword{internal}
