% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_game_data.R
\name{.aggregate_bowling_game_data_cricsheet}
\alias{.aggregate_bowling_game_data_cricsheet}
\title{Aggregate Bowling Stats Per Player Per Match (cricsheet source)}
\usage{
.aggregate_bowling_game_data_cricsheet(
  conn,
  format,
  match_ids = NULL,
  gender = NULL
)
}
\arguments{
\item{conn}{DBI connection}

\item{format}{Character. Match format.}

\item{match_ids}{Character vector or NULL.}

\item{gender}{Character or NULL.}
}
\description{
Aggregate Bowling Stats Per Player Per Match (cricsheet source)
}
\keyword{internal}
