% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_game_data.R
\name{.cricsheet_format_sql}
\alias{.cricsheet_format_sql}
\title{Format filter for cricsheet's match_type vocabulary}
\usage{
.cricsheet_format_sql(column, format)
}
\description{
Format filter for cricsheet's match_type vocabulary
}
\keyword{internal}
