% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_connection.R
\name{.db_resolution}
\alias{.db_resolution}
\title{Read back a resolution tag, defaulting to unknown}
\usage{
.db_resolution(path)
}
\description{
Read back a resolution tag, defaulting to unknown
}
\keyword{internal}
