% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_game_data.R
\name{.join_player_names}
\alias{.join_player_names}
\title{Join Player Names and Team from cricinfo.balls title field}
\usage{
.join_player_names(pgd, conn, format)
}
\arguments{
\item{pgd}{data.table with player_id and match_id columns.}

\item{conn}{DBI connection.}

\item{format}{Character. Match format.}
}
\value{
pgd with added player_name and team columns.
}
\description{
Parses the "Bowler to Batter" title field to extract player names,
then joins team names from cricinfo.matches.
}
\keyword{internal}
