% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_game_data_storage.R
\name{.load_player_game_data_remote}
\alias{.load_player_game_data_remote}
\title{Load Player Game Data from GitHub Release}
\usage{
.load_player_game_data_remote(format, match_ids = NULL, player_ids = NULL)
}
\arguments{
\item{format}{Character. Match format.}

\item{match_ids}{Character vector or NULL.}

\item{player_ids}{Character vector or NULL.}
}
\value{
data.table
}
\description{
Load Player Game Data from GitHub Release
}
\keyword{internal}
