% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_elo.R
\docType{data}
\name{elo_constants}
\alias{elo_constants}
\alias{ELO_START_RATING}
\title{ELO Rating Constants}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ELO_START_RATING
}
\description{
ELO Rating Constants
}
\keyword{internal}
