% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_format_tables.R
\name{elo_params_match}
\alias{elo_params_match}
\title{Compare ELO Parameters}
\usage{
elo_params_match(current, stored)
}
\arguments{
\item{current}{Named list of current parameters}

\item{stored}{Named list of stored parameters}
}
\value{
Logical. TRUE if parameters match, FALSE otherwise
}
\description{
Compares current parameters to stored parameters.
}
\keyword{internal}
