% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality_storage.R
\name{ensure_metric_history_table}
\alias{ensure_metric_history_table}
\title{Ensure a metric history table exists}
\usage{
ensure_metric_history_table(metric, format, conn, gender = NULL)
}
\description{
Ensure a metric history table exists
}
\keyword{internal}
