% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ingestion.R
\name{fast_rbind}
\alias{fast_rbind}
\title{Fast Row Bind}
\usage{
fast_rbind(df_list)
}
\arguments{
\item{df_list}{List of data frames to combine}
}
\value{
Combined data frame
}
\description{
Combines a list of data frames efficiently using data.table::rbindlist.
}
\keyword{internal}
