% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFiltering.R
\name{featureFiltering}
\alias{featureFiltering}
\title{Feature space reduction via filtering methods}
\usage{
featureFiltering(data, target, index = NULL, method = "cc",
  returning = "names")
}
\arguments{
\item{data}{a dataset containing the target, all features, and potentially a time index}

\item{target}{a string with the column name of the target}

\item{index}{a string with the column name of the index}

\item{method}{a string indicating the method used for filtering: "cc", "mrmr"}

\item{returning}{a string indicating the return: "names", "data"}
}
\value{
Either a vector of names of a data.frame object depending on the returning argument
}
\description{
This function .....
}
\details{
some details...
}
\examples{
\dontrun{
# Simulate Data
test_df <- sim_data()

# Correlation Collinearity Filter
test_cc <- featureFiltering(data = test_df,
                            target = "y",
                            method = "cc",
                            returning = "names")

# Maximum Relevance Minimum Redundancy Filter
test_mr <- featureFiltering(data = test_df,
                            target = "y",
                            method = "mrmr",
                            returning = "names")
}
}
