% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{featureSelection}
\alias{featureSelection}
\title{Automated Feature Selection for Machine Learning Models with Boosting Ensembles}
\usage{
featureSelection(data, target, index = NULL, selection = selectionControl(),
  bootstrap = "none", boosting = boostingControl(),
  early_stopping = "none", n_cores = 1, verbose = TRUE)
}
\arguments{
\item{data}{a dataset containing the target, all features, and potentially a time index}

\item{target}{a string with the column name of the target}

\item{index}{a string with the column name of the index}

\item{selection}{takes the return from \code{\link{selectionControl}}}

\item{bootstrap}{a string specifying the bootstrap method}

\item{boosting}{takes the return form \code{\link{boostingControl}}}

\item{early_stopping}{a string giving the method for early stopping}

\item{n_cores}{an integer giving the amount of cores}

\item{verbose}{a logical indicating whether logging info is displayed}
}
\value{
a list with the following entries
\itemize{
\item \code{stability} - a stability matrix with two columns: feature names and selection stability score
\item \code{opt_formula} - statistically optimal formula derived from breakpoints in the stability score series
\item \code{setup} - a list with misc. configurations
}
}
\description{
Automated Feature Selection for Machine Learning Models with Boosting Ensembles
}
\details{
SOME DETAILS
}
\examples{
\dontrun{
# Simulate Data
test_df <- sim_data()

# Genetic Boosting Ensemble
test_ge <- featureSelection(data = test_df,
                            target = "y",
                            method = "geneticboost",
                            selection = selectionControl(n_rounds = 10,
                                                         n_mods = 100,
                                                         p = 30,
                                                         penalty = 0.3,
                                                         reward = 0.2),
                            bootstrap = "regular",
                            early_stopping = "none",
                            n_cores = 1)
}
}
