% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality_core.R
\name{find_bipartite_components}
\alias{find_bipartite_components}
\title{Find Connected Components in Bipartite Graph}
\usage{
find_bipartite_components(matchup_matrix)
}
\arguments{
\item{matchup_matrix}{Matrix of delivery counts (batter x bowler).}
}
\value{
List with:
\itemize{
\item batter_component: Named vector mapping batter_id to component ID
\item bowler_component: Named vector mapping bowler_id to component ID
\item component_sizes: Named vector of component sizes (total players)
\item n_components: Number of connected components found
}
}
\description{
Detects connected components in the batter-bowler matchup graph using
a union-find algorithm. Players in isolated clusters (small components)
will have their PageRank penalized proportionally to component size.
}
\keyword{internal}
