% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_connection.R
\name{force_close_duckdb}
\alias{force_close_duckdb}
\title{Force Close All DuckDB Connections}
\usage{
force_close_duckdb()
}
\value{
Invisibly returns TRUE
}
\description{
Emergency function to release file locks when you get
"file is being used by another process" errors.
}
\details{
This function:
\enumerate{
\item Forces garbage collection to finalize orphaned connections
\item Attempts to shutdown any DuckDB driver instances
}

If the file is still locked after calling this, restart R with Ctrl+Shift+F10.
}
\examples{
\dontrun{
# If you get a file lock error:
force_close_duckdb()

# Then retry your operation
install_all_bouncer_data(fresh = TRUE)
}
}
