% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_scraper.R
\name{fox_combine_matches}
\alias{fox_combine_matches}
\title{Combine all downloaded Fox Sports match files}
\usage{
fox_combine_matches(output_dir = "../bouncerdata/fox_cricket")
}
\arguments{
\item{output_dir}{Directory containing .rds match files (default: bouncerdata/fox_cricket)}
}
\value{
Combined data.frame of all matches
}
\description{
Combine all downloaded Fox Sports match files
}
