% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_scraper.R
\name{fox_match_exists}
\alias{fox_match_exists}
\title{Check if a Fox Sports match ID exists}
\usage{
fox_match_exists(browser, match_id, userkey)
}
\arguments{
\item{browser}{A ChromoteSession object}

\item{match_id}{Fox Sports match ID to check}

\item{userkey}{Valid userkey}
}
\value{
Logical TRUE if match has data, FALSE otherwise
}
\description{
Check if a Fox Sports match ID exists
}
