% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_predictions.R
\name{generate_prediction_report}
\alias{generate_prediction_report}
\title{Generate Prediction Report}
\usage{
generate_prediction_report(prediction)
}
\arguments{
\item{prediction}{List. Prediction from predict_match_outcome}
}
\value{
Character string with formatted report
}
\description{
Creates a formatted prediction report for a match.
}
