% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_boost.R
\name{geneticBoost}
\alias{geneticBoost}
\title{Genetic Boosting Core Algorithm for Feature Selection}
\usage{
geneticBoost(iteration, data, target, selection, bootstrap, early_stopping,
  boosting)
}
\arguments{
\item{iteration}{an integer specifying the iteration}

\item{data}{a dataset containing the target, all features, and potentially a time index}

\item{target}{a string with the column name of the target}

\item{selection}{takes the return from \code{\link{selectionControl}}}

\item{bootstrap}{a string specifying the bootstrap method}

\item{early_stopping}{a string giving the method for early stopping}

\item{boosting}{takes the return form \code{\link{boostingControl}}}
}
\value{
an updated feature Matrix
}
\description{
Genetic Boosting Core Algorithm for Feature Selection
}
