% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_core.R
\name{get_avg_runs_per_ball}
\alias{get_avg_runs_per_ball}
\title{Get Average Runs Per Ball}
\usage{
get_avg_runs_per_ball(match_type)
}
\arguments{
\item{match_type}{Character. Match type}
}
\value{
Numeric. Average runs per ball
}
\description{
Returns the average runs per ball for a given match format.
}
\keyword{internal}
