% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_core.R
\name{get_base_wicket_prob}
\alias{get_base_wicket_prob}
\title{Get Base Wicket Probability}
\usage{
get_base_wicket_prob(match_type)
}
\arguments{
\item{match_type}{Character. Match type}
}
\value{
Numeric. Base wicket probability
}
\description{
Returns the base probability of wicket per delivery for a given format.
}
\keyword{internal}
