% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loaders.R
\name{get_categories_to_load}
\alias{get_categories_to_load}
\title{Get Categories to Load}
\usage{
get_categories_to_load(format = "all", gender = "all", type = "all")
}
\arguments{
\item{format}{"long_form", "short_form", or "all"}

\item{gender}{"male", "female", or "all"}

\item{type}{"international", "club", or "all"}
}
\value{
Character vector of category names.
}
\description{
Returns the category strings based on filter parameters.
}
\keyword{internal}
