% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{get_completed_matches}
\alias{get_completed_matches}
\title{Tidy metadata from completed Test, ODI or IPL T20 matches}
\usage{
get_completed_matches(type = "TEST", competition = "MEN")
}
\arguments{
\item{type}{Type of match data to return: must be one of "TEST", "ODI" or "IPL". Default is "TEST".}

\item{competition}{Men's or Women's competitions: must be one of "MEN" or "WOMEN". Default is "MEN".}
}
\value{
A tibble containing metadata for Test, ODI or IPL T20 matches.
}
\description{
Tidy metadata from completed Test, ODI or IPL T20 matches
}
\examples{
df <- get_completed_matches(type = "TEST", competition = "MEN")
df <- get_completed_matches(type = "ODI", competition = "WOMEN")
df <- get_completed_matches(type = "IPL", competition = "MEN")
}
