% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cricinfo_data.R
\name{get_cricinfo_release}
\alias{get_cricinfo_release}
\title{Get Cricinfo Release Info}
\usage{
get_cricinfo_release()
}
\value{
List with release info (tag_name, assets, etc.).
}
\description{
Gets the cricinfo-rich release info, with session-level caching.
}
\keyword{internal}
