% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_match_prediction.R
\name{get_default_venue_stats}
\alias{get_default_venue_stats}
\title{Get Default Venue Statistics}
\usage{
get_default_venue_stats(format)
}
\arguments{
\item{format}{Character. Match format.}
}
\value{
List with venue statistics.
}
\description{
Returns average venue statistics for a format when specific venue data
is not available.
}
\keyword{internal}
