% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_ids.R
\name{get_format_from_team_id}
\alias{get_format_from_team_id}
\title{Extract Format from Team ID}
\usage{
get_format_from_team_id(team_id)
}
\arguments{
\item{team_id}{Character. Composite team ID}
}
\value{
Character. Format component (t20, odi, test)
}
\description{
Quick extraction of format component from team ID.
}
