% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_3way.R
\name{get_format_gender_suffix}
\alias{get_format_gender_suffix}
\title{Get format-gender suffix for constant lookup}
\usage{
get_format_gender_suffix(format, gender = "male")
}
\arguments{
\item{format}{Format (T20, ODI, TEST, IT20, etc.)}

\item{gender}{Gender (male or female)}
}
\value{
Suffix like "MENS_T20" or "WOMENS_ODI"
}
\description{
Get format-gender suffix for constant lookup
}
\keyword{internal}
