% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_utils.R
\name{get_format_groups}
\alias{get_format_groups}
\title{Get Format Groups Mapping}
\usage{
get_format_groups()
}
\value{
Named list. Keys are canonical formats ("t20", "odi", "test"),
values are character vectors of match types.
}
\description{
Returns a list mapping canonical format names to their database match types.
This consolidates the FORMAT_GROUPS definition that was repeated in 16+ scripts.
}
\examples{
get_format_groups()
# Returns: list(t20 = c("T20", "IT20"), odi = c("ODI", "ODM"), test = c("Test", "MDM"))

get_format_groups()$t20
# Returns: c("T20", "IT20")

}
