% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_data.R
\name{get_fox_release}
\alias{get_fox_release}
\title{Get Fox Sports Release Info}
\usage{
get_fox_release()
}
\value{
List with release info (tag_name, assets, etc.).
}
\description{
Gets the foxsports release info, with session-level caching.
}
\keyword{internal}
