% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkeye.R
\name{get_hawkeye_stats}
\alias{get_hawkeye_stats}
\title{Hawkeye Stats for Tests, ODI or IPL matches}
\usage{
get_hawkeye_stats(id)
}
\arguments{
\item{id}{Id for match(es) to collect Hawkeye stats from. Invalid ids or ids with no Hawkeye data will result in an empty tibble
or an error. Ids can be found using the get_completed_matches function or via the icc-cricket.com results page for Test and
ODI matches and via the iplt20.com page for IPL matches.}
}
\value{
A tibble containing the Hawkey stats for the requested matches. The Hawkeye stats include the ball speed in metres/second,
the x, y location when the ball impacts the pitch, the x, y location of the ball as it passes the stumps and the x, y location
on the field after the batter has played their shot. x, y locations are presented in metres. Individual delivery Hawkeye stats
may be missing if the technology failed. Ball tracking is not available for all matches and field x, y locations are not
available for all matches. Some matches include field x, y locations for non scoring shots, while others only include
field x, y locations for scoring shots.
}
\description{
Hawkeye Stats for Tests, ODI or IPL matches
}
\examples{
df <- get_hawkeye_stats(id = 23469)
}
