% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkeye_all.R
\name{get_hawkeye_stats_all}
\alias{get_hawkeye_stats_all}
\title{Hawkeye Stats for All Tests, ODI or IPL matches}
\usage{
get_hawkeye_stats_all(type = "TEST", competition = "MEN")
}
\arguments{
\item{type}{Type of match data to return: must be one of "TEST", "ODI" or "IPL". Default is "TEST".}

\item{competition}{Men's or Women's competitions: must be one of "MEN" or "WOMEN". Default is "MEN".}
}
\value{
A tibble containing the Hawkey stats for all available matches. The Hawkeye stats include the ball speed in metres/second,
the x, y location when the ball impacts the pitch, the x, y location of the ball as it passes the stumps and the x, y location
on the field after the batter has played their shot. x, y locations are presented in metres. Individual delivery Hawkeye stats
may be missing if the technology failed. Ball tracking is not available for all matches and field x, y locations are not
available for all matches. Some matches include field x, y locations for non scoring shots, while others only include
field x, y locations for scoring shots.
}
\description{
Hawkeye Stats for All Tests, ODI or IPL matches
}
\examples{
df <- get_hawkeye_stats_all(type = "TEST", competition = "WOMEN")
}
