% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loaders.R
\name{get_local_parquet_dir}
\alias{get_local_parquet_dir}
\title{Get Local Parquet Directory}
\usage{
get_local_parquet_dir()
}
\value{
Path to local parquet directory.
}
\description{
Returns the path to local parquet files.
}
\keyword{internal}
