% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_utils.R
\name{get_match_types_for_format}
\alias{get_match_types_for_format}
\title{Get Match Types for Format}
\usage{
get_match_types_for_format(format)
}
\arguments{
\item{format}{Character. Canonical format ("t20", "odi", "test") or
match type that will be normalized first.}
}
\value{
Character vector. Database match types for the format.
}
\description{
Returns the database match type values for a given canonical format.
}
\examples{
get_match_types_for_format("t20")   # c("T20", "IT20")
get_match_types_for_format("Test")  # c("Test", "MDM")

}
