% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_utils.R
\name{get_max_balls}
\alias{get_max_balls}
\title{Get Maximum Balls for Format}
\usage{
get_max_balls(format)
}
\arguments{
\item{format}{Character. Cricket format ("t20", "odi", "test").}
}
\value{
Integer. Maximum balls: 120 (T20), 300 (ODI), 540 (Test).
}
\description{
Returns the maximum number of balls in an innings for a given format.
}
\examples{
get_max_balls("t20")   # 120
get_max_balls("odi")   # 300
get_max_balls("test")  # 540

}
