% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_utils.R
\name{get_max_overs}
\alias{get_max_overs}
\title{Get Maximum Overs for Format}
\usage{
get_max_overs(format)
}
\arguments{
\item{format}{Character. Cricket format ("t20", "odi", "test").}
}
\value{
Integer or NULL. Maximum overs: 20 (T20), 50 (ODI), NULL (Test unlimited).
}
\description{
Returns the maximum number of overs in an innings for a given format.
}
\examples{
get_max_overs("t20")   # 20
get_max_overs("odi")   # 50
get_max_overs("test")  # NULL

}
