% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_outcomes.R
\name{get_models_dir}
\alias{get_models_dir}
\title{Get Models Directory}
\usage{
get_models_dir()
}
\value{
Character string with path to models directory
}
\description{
Finds the models directory within the bouncerdata directory.
}
\keyword{internal}
