% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_elo_dynamic.R
\name{get_player_tier_starting_elo}
\alias{get_player_tier_starting_elo}
\title{Get Player Tier Starting ELO}
\usage{
get_player_tier_starting_elo(event_name)
}
\arguments{
\item{event_name}{Character. The event/tournament name.}
}
\value{
Numeric. Starting ELO for the player (1400-1525).
}
\description{
Returns the starting ELO for a new player based on the tier of the event
they first appear in. Players debuting in elite events (IPL, BBL) start
higher than players debuting in development tournaments.
}
\details{
This leverages the existing event tier system from team ELO.
}
\examples{
get_player_tier_starting_elo("Indian Premier League")  # 1525
get_player_tier_starting_elo("Vitality Blast")         # 1500
get_player_tier_starting_elo("ACC Qualifier")          # 1450
get_player_tier_starting_elo("Kwibuka T20")            # 1400
}
