% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_match_features.R
\name{get_prediction_feature_cols}
\alias{get_prediction_feature_cols}
\title{Get Prediction Feature Columns}
\usage{
get_prediction_feature_cols()
}
\value{
Character vector of feature column names
}
\description{
Returns the list of feature column names used for XGBoost prediction models.
}
\keyword{internal}
