% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_3way.R
\name{get_runs_per_100_elo}
\alias{get_runs_per_100_elo}
\title{Get runs per 100 ELO points for format-gender}
\usage{
get_runs_per_100_elo(format, gender = "male")
}
\arguments{
\item{format}{Format (T20, ODI, TEST)}

\item{gender}{Gender (male or female)}
}
\value{
Numeric value
}
\description{
Get runs per 100 ELO points for format-gender
}
