% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_skill.R
\name{get_skill_alpha_params}
\alias{get_skill_alpha_params}
\title{Get Skill Index Alpha (Learning Rate) Parameters}
\usage{
get_skill_alpha_params(format, gender = "male", skill_type = "run")
}
\arguments{
\item{format}{Format (T20, ODI, TEST)}

\item{gender}{Gender (male or female)}

\item{skill_type}{"run" or "wicket"}
}
\value{
Named list with alpha_max, alpha_min, halflife
}
\description{
Get Skill Index Alpha (Learning Rate) Parameters
}
\examples{
get_skill_alpha_params("t20")
get_skill_alpha_params("odi", skill_type = "wicket")

}
