% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skill_utils.R
\name{get_skill_table_name}
\alias{get_skill_table_name}
\title{Get Format-Specific Skill Table Name}
\usage{
get_skill_table_name(format, entity_type)
}
\arguments{
\item{format}{Character. Format name (will be normalized)}

\item{entity_type}{Character. One of "player_skill", "team_skill", "venue_skill"}
}
\value{
Character. Table name like "t20_player_skill"
}
\description{
Returns the database table name for a format and entity type.
}
\keyword{internal}
