% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_skill.R
\name{get_skill_weights}
\alias{get_skill_weights}
\title{Get Skill Index Attribution Weights}
\usage{
get_skill_weights(format, gender = "male", skill_type = "run")
}
\arguments{
\item{format}{Format (T20, ODI, TEST)}

\item{gender}{Gender (male or female)}

\item{skill_type}{"run" or "wicket"}
}
\value{
Named list with w_batter, w_bowler, w_venue_session, w_venue_perm
}
\description{
Get Skill Index Attribution Weights
}
