% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_way_elo.R
\name{get_stored_3way_elo_params}
\alias{get_stored_3way_elo_params}
\title{Get Stored 3-Way ELO Parameters}
\usage{
get_stored_3way_elo_params(format, conn)
}
\arguments{
\item{format}{Character. Format identifier.}

\item{conn}{A DuckDB connection object.}
}
\value{
List with parameters or NULL if not found.
}
\description{
Retrieves stored parameters from database.
}
\keyword{internal}
