% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_match_features.R
\name{get_team_elo_fast}
\alias{get_team_elo_fast}
\title{Get Team ELO (Fast Version)}
\usage{
get_team_elo_fast(team, as_of_date, team_elo_dt)
}
\arguments{
\item{team}{Character. Team ID (composite format: team_gender_format)}

\item{as_of_date}{Date. Get ELO as of this date}

\item{team_elo_dt}{data.table. Pre-loaded team ELO data}
}
\value{
List with elo_result and elo_roster
}
\description{
Fast lookup of team ELO from pre-loaded data.table.
}
