% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_match_features.R
\name{get_team_skills_fast}
\alias{get_team_skills_fast}
\title{Get Team Skills (Fast Version)}
\usage{
get_team_skills_fast(
  team,
  as_of_date,
  player_participation_dt,
  batter_skills,
  bowler_skills,
  n_recent = 3
)
}
\arguments{
\item{team}{Character. Team name (not composite ID)}

\item{as_of_date}{Date. Calculate skills as of this date}

\item{player_participation_dt}{data.table. Pre-loaded player participation data}

\item{batter_skills}{data.table. Pre-loaded batter skill indices}

\item{bowler_skills}{data.table. Pre-loaded bowler skill indices}

\item{n_recent}{Integer. Number of recent matches to infer roster (default 3)}
}
\value{
List with batting and bowling skill aggregates
}
\description{
Fast calculation of team skill aggregates from pre-loaded data.
}
\keyword{internal}
