% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_3way.R
\name{get_venue_k_factors}
\alias{get_venue_k_factors}
\title{Get venue K-factors for format-gender}
\usage{
get_venue_k_factors(format, gender = "male")
}
\arguments{
\item{format}{Format (T20, ODI, TEST)}

\item{gender}{Gender (male or female)}
}
\value{
Named list with perm and session k-factor parameters
}
\description{
Get venue K-factors for format-gender
}
\examples{
get_venue_k_factors("t20")
get_venue_k_factors("test")

}
