% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skill_indices.R
\name{get_venue_session_skill_alpha}
\alias{get_venue_session_skill_alpha}
\title{Get Venue Session Skill Alpha}
\usage{
get_venue_session_skill_alpha(balls_in_match)
}
\arguments{
\item{balls_in_match}{Integer. Balls bowled in current match.}
}
\value{
Numeric. The alpha for session venue skill.
}
\description{
Calculates the learning rate for session venue skill.
Session skill learns fast within a match to capture current conditions.
}
\keyword{internal}
