% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_3way.R
\name{get_wicket_elo_divisor}
\alias{get_wicket_elo_divisor}
\title{Get wicket ELO divisor for format-gender}
\usage{
get_wicket_elo_divisor(format, gender = "male")
}
\arguments{
\item{format}{Format (T20, ODI, TEST)}

\item{gender}{Gender (male or female)}
}
\value{
Numeric value (400 or 200)
}
\description{
Get wicket ELO divisor for format-gender
}
\examples{
get_wicket_elo_divisor("t20")
get_wicket_elo_divisor("t20", gender = "female")

}
