% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_3way.R
\name{get_wicket_elo_weights}
\alias{get_wicket_elo_weights}
\title{Get Wicket ELO attribution weights for format-gender}
\usage{
get_wicket_elo_weights(format, gender = "male")
}
\arguments{
\item{format}{Format (T20, ODI, TEST)}

\item{gender}{Gender (male or female)}
}
\value{
Named list with w_batter, w_bowler, w_venue_session, w_venue_perm
}
\description{
Get Wicket ELO attribution weights for format-gender
}
