% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_3way.R
\name{get_wicket_k_factors}
\alias{get_wicket_k_factors}
\title{Get player Wicket K-factors for format-gender}
\usage{
get_wicket_k_factors(format, gender = "male")
}
\arguments{
\item{format}{Format (T20, ODI, TEST)}

\item{gender}{Gender (male or female)}
}
\value{
Named list with k_max, k_min, halflife
}
\description{
Get player Wicket K-factors for format-gender
}
