% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ingestion.R
\name{has_datatable_support}
\alias{has_datatable_support}
\title{Check for data.table Support}
\usage{
has_datatable_support()
}
\value{
Logical. TRUE if data.table is available.
}
\description{
Checks if the data.table package is available for fast data frame binding.
data.table::rbindlist is ~8x faster than do.call(rbind, ...) for combining
many data frames.
}
\keyword{internal}
