% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_outcomes.R
\name{identify_super_over_matches}
\alias{identify_super_over_matches}
\title{Identify Super Over Matches}
\usage{
identify_super_over_matches(conn, event_filter = NULL, match_type = "t20")
}
\arguments{
\item{conn}{DuckDB connection}

\item{event_filter}{Character. Event name filter. Default NULL = all.}

\item{match_type}{Character. Match type filter. Default "t20".}
}
\value{
Data frame with super over match information
}
\description{
Queries the database to identify matches decided by super over.
}
\examples{
\dontrun{
conn <- get_db_connection()
super_overs <- identify_super_over_matches(conn, event_filter = "IPL")
DBI::dbDisconnect(conn, shutdown = TRUE)
}
}
