% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_processing.R
\name{initialize_player_state}
\alias{initialize_player_state}
\title{Initialize Player State}
\usage{
initialize_player_state(conn, fresh, matches_with_elo)
}
\description{
Initialize Player State
}
\keyword{internal}
