% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_correlation.R
\name{jaccard_similarity}
\alias{jaccard_similarity}
\title{Calculate Jaccard Similarity}
\usage{
jaccard_similarity(set_a, set_b)
}
\arguments{
\item{set_a}{First set (vector)}

\item{set_b}{Second set (vector)}
}
\value{
Numeric between 0 and 1
}
\description{
Computes Jaccard index between two sets: |A ∩ B| / |A ∪ B|
}
\keyword{internal}
