% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cricinfo_data.R
\name{load_cricinfo_innings_remote}
\alias{load_cricinfo_innings_remote}
\title{Load Cricinfo Innings from Remote}
\usage{
load_cricinfo_innings_remote(match_ids, format, gender)
}
\arguments{
\item{match_ids}{Character vector of match IDs (required for remote).}

\item{format}{Character. Format filter.}

\item{gender}{Character. Gender filter.}
}
\value{
Data frame.
}
\description{
Load Cricinfo Innings from Remote
}
\keyword{internal}
