% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_processing.R
\name{load_deliveries_to_process}
\alias{load_deliveries_to_process}
\title{Load Deliveries to Process}
\usage{
load_deliveries_to_process(conn, fresh, match_limit)
}
\description{
Load Deliveries to Process
}
\keyword{internal}
