% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_ratings.R
\name{load_stat_ratings}
\alias{load_stat_ratings}
\title{Load Stat Ratings}
\usage{
load_stat_ratings(
  format = c("t20", "odi", "test"),
  source = c("local", "remote"),
  path = NULL
)
}
\arguments{
\item{format}{Character. Match format.}

\item{source}{Character. "local" or "remote".}

\item{path}{Character. Custom DB path (local only).}
}
\value{
data.table with player stat ratings.
}
\description{
Load Stat Ratings
}
