% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{load_venue_coordinates}
\alias{load_venue_coordinates}
\title{Load Venue Coordinates from Database}
\usage{
load_venue_coordinates(conn = NULL)
}
\arguments{
\item{conn}{DBI connection. If NULL, opens one.}
}
\value{
data.table with venue, latitude, longitude.
}
\description{
Load Venue Coordinates from Database
}
