% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_ids.R
\name{make_team_id_vec}
\alias{make_team_id_vec}
\title{Create Composite Team ID (Vectorized)}
\usage{
make_team_id_vec(team, gender, format, team_type)
}
\arguments{
\item{team}{Character vector. Team names}

\item{gender}{Character vector. Genders}

\item{format}{Character vector or single value. Match formats}

\item{team_type}{Character vector. Team types}
}
\value{
Character vector. Composite team IDs in snake_case
}
\description{
Vectorized version of make_team_id for use with data.table/dplyr.
}
\examples{
# In data.table:
# dt[, team_id := make_team_id_vec(team, gender, format, team_type)]
}
