% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{mbb}
\alias{mbb}
\title{A function which computes a moving block bootstrap of a target time series and its respective features.}
\usage{
mbb(data, target, seed = 1)
}
\arguments{
\item{data}{a dataset containing the target time series and its features}

\item{target}{a character with the target variable}

\item{seed}{a numeric handling the random generator}
}
\value{
BLA BLA
}
\description{
A function which computes a moving block bootstrap of a target time series and its respective features.
}
\examples{
# Simulate test data
test_data <- sim_data()
# Bootstrap
bootstrapped <- mbb(data = test_data,
                    target = "y",
                    seed = 1)
}
