% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_utils.R
\name{normalize_match_type}
\alias{normalize_match_type}
\title{Normalize Match Type (Deprecated)}
\usage{
normalize_match_type(match_type)
}
\arguments{
\item{match_type}{Character. Match type to normalize}
}
\value{
Character. Canonical format.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Use \code{\link[=normalize_format]{normalize_format()}} instead.
}
\keyword{internal}
