% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_scraper.R
\name{parse_fox_aggregatestats}
\alias{parse_fox_aggregatestats}
\title{Parse Fox Sports aggregatestats response to data.frame}
\usage{
parse_fox_aggregatestats(json_data, match_id)
}
\arguments{
\item{json_data}{List from Fox Sports API response}

\item{match_id}{Character string of the match ID}
}
\value{
A data.frame of ball-by-ball data, or NULL if no data
}
\description{
Parse Fox Sports aggregatestats response to data.frame
}
\keyword{internal}
