% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_scraper.R
\name{parse_fox_details}
\alias{parse_fox_details}
\title{Parse Fox Sports details.json response to data.frame}
\usage{
parse_fox_details(json_data, match_id)
}
\arguments{
\item{json_data}{List from Fox Sports API details response}

\item{match_id}{Character string of the match ID}
}
\value{
A single-row data.frame of match details, or NULL if no data
}
\description{
Parse Fox Sports details.json response to data.frame
}
\keyword{internal}
