% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_scraper.R
\name{parse_fox_players}
\alias{parse_fox_players}
\title{Parse Fox Sports players.json response to data.frame}
\usage{
parse_fox_players(json_data, match_id)
}
\arguments{
\item{json_data}{List from Fox Sports API players response}

\item{match_id}{Character string of the match ID}
}
\value{
A data.frame of player data for both teams, or NULL if no data
}
\description{
Parse Fox Sports players.json response to data.frame
}
\keyword{internal}
