% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cricsheet_parser.R
\name{parse_match_info}
\alias{parse_match_info}
\title{Parse Match Info}
\usage{
parse_match_info(json_data, match_id)
}
\arguments{
\item{json_data}{Parsed JSON data}

\item{match_id}{Match ID (from filename)}
}
\value{
A data frame with match information
}
\description{
Extracts match metadata from Cricsheet JSON.
}
\keyword{internal}
