% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_predictions.R
\name{predict_from_elo}
\alias{predict_from_elo}
\title{Predict from ELO}
\usage{
predict_from_elo(features)
}
\arguments{
\item{features}{Data frame. Pre-match features}
}
\value{
Numeric. Team 1 win probability
}
\description{
Simple prediction based on ELO difference when model is not available.
}
\keyword{internal}
