% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_match_prediction.R
\name{predict_with_features}
\alias{predict_with_features}
\title{Predict using XGBoost model with feature alignment}
\usage{
predict_with_features(model, data, feature_cols)
}
\arguments{
\item{model}{XGBoost model}

\item{data}{data.frame with feature data}

\item{feature_cols}{Character vector of expected feature columns}
}
\value{
Numeric prediction (0-1)
}
\description{
Predict using XGBoost model with feature alignment
}
\keyword{internal}
