% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_outcomes.R
\name{prepare_longform_features}
\alias{prepare_longform_features}
\title{Prepare Long-form Features for Prediction}
\usage{
prepare_longform_features(df)
}
\description{
Prepare Long-form Features for Prediction
}
\keyword{internal}
