% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_outcomes.R
\name{prepare_shortform_features}
\alias{prepare_shortform_features}
\title{Prepare Short-form Features for Prediction}
\usage{
prepare_shortform_features(df)
}
\description{
Prepare Short-form Features for Prediction
}
\keyword{internal}
