% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_match_prediction.R
\name{print.bouncer_win_prob}
\alias{print.bouncer_win_prob}
\title{Print method for bouncer_win_prob objects}
\usage{
\method{print}{bouncer_win_prob}(x, ...)
}
\arguments{
\item{x}{A bouncer_win_prob object from predict_win_probability()}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for bouncer_win_prob objects
}
