% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_audit.R
\name{print.low_information_audit}
\alias{print.low_information_audit}
\title{Print method for low_information_audit objects}
\usage{
\method{print}{low_information_audit}(x, ...)
}
\arguments{
\item{x}{A \code{low_information_audit} object from \code{\link[=audit_low_information_state]{audit_low_information_state()}}.}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for low_information_audit objects
}
