% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_validation.R
\docType{data}
\name{RATING_VAL_MIN_CAREER_BALLS}
\alias{RATING_VAL_MIN_CAREER_BALLS}
\title{Minimum career balls before the origin for a player to be evaluated}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
RATING_VAL_MIN_CAREER_BALLS
}
\description{
Minimum career balls before the origin for a player to be evaluated
}
\keyword{datasets}
