% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_validation.R
\docType{data}
\name{RATING_VAL_MIN_WINDOW_BALLS}
\alias{RATING_VAL_MIN_WINDOW_BALLS}
\title{Minimum balls inside the 12-month scoring window}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
RATING_VAL_MIN_WINDOW_BALLS
}
\description{
Minimum balls inside the 12-month scoring window
}
\keyword{datasets}
