% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_validation.R
\docType{data}
\name{RATING_VAL_ORIGINS}
\alias{RATING_VAL_ORIGINS}
\title{Default evaluation origins for the validation harness}
\format{
An object of class \code{Date} of length 10.
}
\usage{
RATING_VAL_ORIGINS
}
\description{
1 January of each year from 2016 to 2025. 2016 is calibration-only (there is no
prior origin to fit on), so scored origins are 2017-2025.
}
\keyword{datasets}
