% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\docType{data}
\name{RUN_RATE_PRIOR_BALLS}
\alias{RUN_RATE_PRIOR_BALLS}
\title{Shrinkage weight for the in-progress run rate, in balls}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
RUN_RATE_PRIOR_BALLS
}
\description{
Shrinkage weight for the in-progress run rate, in balls
}
\keyword{internal}
