% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\docType{data}
\name{RUN_RATE_PRIOR_RATE}
\alias{RUN_RATE_PRIOR_RATE}
\title{Prior run rate per format, runs per over}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
RUN_RATE_PRIOR_RATE
}
\description{
The empirical mean innings rate over all innings of 60+ balls: T20 7.252,
ODI 4.996, Test 3.343.
}
\keyword{internal}
