% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{save_match_weather}
\alias{save_match_weather}
\title{Save Match Weather Data to Database}
\usage{
save_match_weather(weather, conn = NULL)
}
\arguments{
\item{weather}{data.frame with match weather data.}

\item{conn}{DBI connection. If NULL, opens one.}
}
\description{
Save Match Weather Data to Database
}
\keyword{internal}
