% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_projection_stage2.R
\name{save_test_stage2_correction}
\alias{save_test_stage2_correction}
\alias{load_test_stage2_correction}
\title{Save/load a fitted stage-2 correction (mirrors save/load_test_overs_model())}
\usage{
save_test_stage2_correction(fit_result, models_dir = NULL)

load_test_stage2_correction(match_type = c("Test", "MDM"), models_dir = NULL)
}
\description{
Save/load a fitted stage-2 correction (mirrors save/load_test_overs_model())
}
\keyword{internal}
