% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{save_venue_coordinates}
\alias{save_venue_coordinates}
\title{Save Venue Coordinates to Database}
\usage{
save_venue_coordinates(coords, conn = NULL)
}
\arguments{
\item{coords}{data.frame with venue, latitude, longitude columns.}

\item{conn}{DBI connection. If NULL, opens one.}
}
\description{
Save Venue Coordinates to Database
}
\keyword{internal}
