% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\docType{class}
\name{sel_obj-class}
\alias{sel_obj-class}
\alias{plot,sel_obj,ANY-method}
\alias{summary,sel_obj-method}
\alias{print,sel_obj-method}
\title{A S4 class to represent the a feature selection output obtained by the \code{\link{featureSelection}} function.}
\usage{
\S4method{plot}{sel_obj,ANY}(x, n_features = NULL)

\S4method{summary}{sel_obj}(object)

\S4method{print}{sel_obj}(x)
}
\arguments{
\item{x}{an object of class \linkS4class{sel_obj}}

\item{n_features}{a numeric value determining the number of variables in the plot}

\item{object}{an object of class \linkS4class{sel_obj}}
}
\description{
A S4 class to represent the a feature selection output obtained by the \code{\link{featureSelection}} function.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: A variable importance plot

\item \code{summary}: A summary of an output obtained by \code{\link{featureSelection}}

\item \code{print}: A printed setup of a \linkS4class{sel_obj}
}}

\section{Slots}{

\describe{
\item{\code{stability}}{a \code{\link{tibble}} with the stability matrix obtained by the feature selection algorithm}

\item{\code{opt_formula}}{a \code{\link{formula}} object as proposed by a structural break test.}

\item{\code{setup}}{a \code{\link{list}} containing call and runtime information.}
}}

