% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionControl.R
\name{selectionControl}
\alias{selectionControl}
\title{Control Function for the Selection Setup}
\usage{
selectionControl(n_rounds = NULL, n_mods = NULL, p = NULL,
  penalty = NULL, reward = NULL)
}
\arguments{
\item{n_rounds}{an integer specifying the number of outer loop iterations}

\item{n_mods}{an integer specifying the number of models inside each genetic mutation state}

\item{p}{an integer specifying the number of features for each weak learner}

\item{penalty}{an integer between 0 and 1 giving the penalty in the genetic setting}

\item{reward}{an integer between 0 and 1 giving the reward in the genetic setting}
}
\value{
a list of the specified input
}
\description{
Control Function for the Selection Setup
}
\examples{
selection = selectionControl(n_rounds = 1000,
                             n_mods = 100)

}
