% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{set_simulation_seed}
\alias{set_simulation_seed}
\title{Set Simulation Seed}
\usage{
set_simulation_seed(seed)
}
\arguments{
\item{seed}{Integer. Random seed}
}
\value{
Invisibly returns the seed
}
\description{
Sets the random seed for reproducible simulations.
}
\keyword{internal}
