% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulating Classification and Regression Data}
\usage{
sim_data(n = 1000, modelvars = 50, noisevars = 100, model_sd = 4,
  noise_sd = 4, epsilon_sd = 4, outcome = "regression", cutoff = NULL)
}
\arguments{
\item{n}{an integer giving the number of observations the simulated dataset shall have}

\item{modelvars}{an integer giving the number of relevant predictors, out of the entire feature space}

\item{noisevars}{an integer giving the number of irrelevant predictors}

\item{model_sd}{an integer giving the standard deviation of the model}

\item{noise_sd}{an integer giving the standard deviation of the noise features}

\item{epsilon_sd}{an integer giving the standard deviation of the error term}

\item{outcome}{a string indicating whether a "regression" or a "classification" dataset are supposed to be simulated}

\item{cutoff}{a numeric value specifying the cutoff value for classification simulation}
}
\value{
a data.frame object
}
\description{
Simulating Classification and Regression Data
}
\examples{
# simulating regression data
sim_reg_data <- sim_data(n = 1000, outcome = "regression")

}
