% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality_storage.R
\name{store_metric_snapshot}
\alias{store_metric_snapshot}
\title{Store a metric snapshot (centrality or pagerank)}
\usage{
store_metric_snapshot(
  result,
  snapshot_date,
  metric,
  format,
  conn,
  gender = NULL,
  verbose = TRUE
)
}
\description{
Store a metric snapshot (centrality or pagerank)
}
\keyword{internal}
