% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_predictions.R
\name{store_prediction}
\alias{store_prediction}
\title{Store Prediction}
\usage{
store_prediction(prediction, conn)
}
\arguments{
\item{prediction}{List. Prediction from predict_match_outcome}

\item{conn}{DBI connection. Database connection}
}
\value{
Invisibly returns TRUE on success
}
\description{
Stores a prediction in the database.
}
\keyword{internal}
