% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_ratings.R
\name{store_stat_ratings}
\alias{store_stat_ratings}
\title{Store Stat Ratings to DuckDB}
\usage{
store_stat_ratings(conn, ratings, format = c("t20", "odi", "test"))
}
\arguments{
\item{conn}{DBI connection (writable).}

\item{ratings}{data.table from \code{\link{estimate_player_stat_ratings}}.}

\item{format}{Character. Match format.}
}
\value{
Number of rows stored (invisibly).
}
\description{
Store Stat Ratings to DuckDB
}
\keyword{internal}
