% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_projection_test_match.R
\name{test_overs_to_balls}
\alias{test_overs_to_balls}
\title{Convert Test Overs Remaining to Balls}
\usage{
test_overs_to_balls(overs_remaining)
}
\arguments{
\item{overs_remaining}{Numeric. Overs remaining in match.}
}
\value{
Integer. Balls remaining.
}
\description{
Converts overs remaining to balls for use in projection functions.
}
\examples{
\dontrun{
test_overs_to_balls(45.3)  # 273 balls (45*6 + 3)
}
}
\keyword{internal}
