% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{theme_bouncer}
\alias{theme_bouncer}
\title{Bouncer ggplot2 Theme}
\usage{
theme_bouncer(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme object.
}
\description{
A clean, consistent theme for bouncer visualizations.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(data, aes(x, y)) + geom_point() + theme_bouncer()
}
}
