% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_way_elo_tables.R
\name{three_way_elo_gender_prefix}
\alias{three_way_elo_gender_prefix}
\title{Gender Prefix Used by the 3-Way ELO Tables}
\usage{
three_way_elo_gender_prefix(gender)
}
\arguments{
\item{gender}{Character. \code{"male"}/\code{"mens"} or \code{"female"}/\code{"womens"}.}
}
\value{
\code{"mens"} or \code{"womens"}.
}
\description{
Gender Prefix Used by the 3-Way ELO Tables
}
\keyword{internal}
