% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_way_elo_tables.R
\name{three_way_elo_table}
\alias{three_way_elo_table}
\title{Name of a 3-Way ELO Table}
\usage{
three_way_elo_table(format, gender)
}
\arguments{
\item{format}{Character. Anything \code{\link[=normalize_format]{normalize_format()}} accepts.}

\item{gender}{Character. See \code{\link[=three_way_elo_gender_prefix]{three_way_elo_gender_prefix()}}.}
}
\value{
A single table name, e.g. \code{"mens_t20_3way_elo"}.
}
\description{
Name of a 3-Way ELO Table
}
\keyword{internal}
