% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cricsheet_parser.R
\name{extract_players}
\alias{extract_players}
\title{Extract Players}
\usage{
extract_players(players_data)
}
\arguments{
\item{players_data}{Players section from JSON info}
}
\value{
Data frame with player information
}
\description{
Extracts player information from JSON with deduplication.
}
\keyword{internal}
