% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_3way.R
\name{get_3way_constant}
\alias{get_3way_constant}
\title{Look up a format-gender-specific 3-way ELO constant}
\usage{
get_3way_constant(prefix, format, gender = "male")
}
\arguments{
\item{prefix}{Constant prefix (e.g., "THREE_WAY_W_BATTER")}

\item{format}{Format (T20, ODI, TEST)}

\item{gender}{Gender (male or female)}
}
\value{
The constant value
}
\description{
Look up a format-gender-specific 3-way ELO constant
}
\keyword{internal}
