% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_elo_optimization.R
\name{get_dynamic_k}
\alias{get_dynamic_k}
\title{Get Dynamic K Factor}
\usage{
get_dynamic_k(matches_played, k_max, k_min, k_halflife)
}
\arguments{
\item{matches_played}{Number of matches the team has played}

\item{k_max}{Maximum K factor for new teams}

\item{k_min}{Minimum K factor floor}

\item{k_halflife}{Matches until K decays halfway}
}
\value{
K factor value
}
\description{
Calculates the K factor for a team based on matches played.
Uses exponential decay from k_max to k_min.
}
\examples{
get_dynamic_k(0, k_max = 40, k_min = 20, k_halflife = 50)
get_dynamic_k(100, k_max = 40, k_min = 20, k_halflife = 50)

}
