% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_simulation_seeds}
\alias{get_simulation_seeds}
\title{Get Simulation Seeds}
\usage{
get_simulation_seeds(n, base_seed = 42)
}
\arguments{
\item{n}{Integer. Number of seeds to generate}

\item{base_seed}{Integer. Base seed (default 42)}
}
\value{
Integer vector of seeds
}
\description{
Generates a vector of seeds for parallel simulations.
}
\keyword{internal}
