\alias{peims-package}
\alias{peims}
\author{Jakob Schöpe (\email{jas@med-imbei.uni-saarland.de})}
\description{The \pkg{peims} package provides an efficient utility routine for assessing instability in model selection with subsequent parameter estimation using resampling methods and bagging methodology.}
\details{}
\docType{package}
\name{peims-package}
\references{
Breiman L (1996) Bagging predictors. Mach Learn 24:123--140

Efron B (2014) Estimation and accuracy after model selection. J Am Stat Assoc 109:991--1007
}
\title{peims: Parameter estimation considering instability in model selection using resampling methods}
