% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_state.R
\name{pipeline_should_run}
\alias{pipeline_should_run}
\title{Check if a pipeline step should run}
\usage{
pipeline_should_run(step, steps_to_run = NULL)
}
\arguments{
\item{step}{Step number to check}

\item{steps_to_run}{Vector of steps to run, or NULL for all}
}
\value{
TRUE if step should run
}
\description{
Check if a pipeline step should run
}
\keyword{internal}
