% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality_storage.R
\name{print_centrality_summary}
\alias{print_centrality_summary}
\title{Print Centrality Summary}
\usage{
print_centrality_summary(centrality_result, n_top = 10)
}
\arguments{
\item{centrality_result}{Result from calculate_player_centrality().}

\item{n_top}{Integer. Number of top players to show per role.}
}
\description{
Prints a formatted summary of network centrality results.
}
