% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality_interface.R
\name{print_pagerank_summary}
\alias{print_pagerank_summary}
\title{Print PageRank Summary}
\usage{
print_pagerank_summary(pagerank_result, n_top = 10)
}
\arguments{
\item{pagerank_result}{Result from calculate_player_pagerank().}

\item{n_top}{Integer. Number of top players to show per role.}
}
\description{
Prints a formatted summary of PageRank results.
}
